+++ packages/ragbits-document-search/src/ragbits/document_search/retrieval/rerankers/answerai.py
@@ -1,5 +1,6 @@
 from collections.abc import Sequence
 from itertools import chain
+from typing import Any
 
 from rerankers import Reranker as AnswerReranker
 
@@ -15,7 +16,12 @@ class AnswerAIReranker(Reranker[RerankerOptions]):
 
     options_cls = RerankerOptions
 
-    def __init__(self, model: str, default_options: RerankerOptions | None = None, **rerankers_kwargs: str) -> None:
+    def __init__(
+        self,
+        model: str,
+        default_options: RerankerOptions | None = None,
+        **rerankers_kwargs: Any,  # noqa: ANN401
+    ) -> None:
         """
         Constructs a new AnswerDotAIRerankersReranker instance.
 
@@ -35,7 +41,7 @@ class AnswerAIReranker(Reranker[RerankerOptions]):
         options: RerankerOptions | None = None,
     ) -> Sequence[Element]:
         """
-        Rerank elements .
+        Rerank elements.
 
         Args:
             elements: The elements to rerank.
@@ -51,8 +57,9 @@ class AnswerAIReranker(Reranker[RerankerOptions]):
             IndexError: Raised if docs is an empty List.
         """
         merged_options = (self.default_options | options) if options else self.default_options
-        element_list = list(chain.from_iterable(elements))
-        documents = [element.text_representation for element in element_list]
+        flat_elements = list(chain.from_iterable(elements))
+        documents = [element.text_representation or "" for element in flat_elements]
+
         with trace(
             query=query, documents=documents, elements=elements, model=self.model, options=merged_options
         ) as outputs:
@@ -62,6 +69,14 @@ class AnswerAIReranker(Reranker[RerankerOptions]):
             )
             if merged_options.top_n:
                 response = response.top_k(merged_options.top_n)
-            outputs.results = [element_list[result.document.doc_id] for result in response]
 
-            return outputs.results
+            results = []
+            for result in response:
+                if not merged_options.score_threshold or result.score >= merged_options.score_threshold:
+                    if merged_options.override_score:
+                        flat_elements[result.document.doc_id].score = result.score
+                    results.append(flat_elements[result.document.doc_id])
+
+            outputs.results = results
+
+        return outputs.results