+++ packages/ragbits-document-search/src/ragbits/document_search/retrieval/rerankers/base.py
@@ -10,11 +10,17 @@ from ragbits.document_search.retrieval import rerankers
 
 class RerankerOptions(Options):
     """
-    Options for the reranker.
+    An object representing the options for the reranker.
+
+    Attributes:
+        top_n: The number of entries to return.
+        score_threshold: The minimum relevance score for an entry to be returned.
+        override_score: If True reranking will override element score.
     """
 
     top_n: int | None = None
-    max_chunks_per_doc: int | None = None
+    score_threshold: float | None = None
+    override_score: bool = True
 
 
 RerankerOptionsT = TypeVar("RerankerOptionsT", bound=RerankerOptions)
@@ -25,8 +31,8 @@ class Reranker(ConfigurableComponent[RerankerOptionsT], ABC):
     Reranks elements retrieved from vector store.
     """
 
-    default_module: ClassVar = rerankers
     options_cls: type[RerankerOptionsT]
+    default_module: ClassVar = rerankers
     configuration_key: ClassVar = "reranker"
 
     @abstractmethod