+++ packages/ragbits-document-search/src/ragbits/document_search/retrieval/rerankers/noop.py
@@ -31,6 +31,4 @@ class NoopReranker(Reranker[RerankerOptions]):
         Returns:
             The reranked elements.
         """
-        element_list = [*{element.id: element for element in chain.from_iterable(elements)}.values()]
-
-        return element_list
+        return [*{element.id: element for element in chain.from_iterable(elements)}.values()]