+++ packages/ragbits-document-search/CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## Unreleased
 
+- Add optional score override on reranking step (#544)
+- Add score threshold to reranker options (#544)
 - Add LLM reranker (#513)
+- fix: fix passing score in DocumentSearch.search (#545)
 
 ## 0.16.0 (2025-04-29)